requirements.txt
opencv-python
numpy
tensorflow
Pillow
customtkinter
pyserial
mysql-connector-python